backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
langchain>=0.3.0
langchain-community>=0.3.0
boto3>=1.34.58
faiss-cpu==1.7.4
sentence-transformers==2.3.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
slowapi==0.1.9
